Rename: PostPageHeader.jsx
- List page 내의 paper와 네이밍 혼선이 있어 용도기준으로 컴포넌트 명 변경
