+++ src/components/Card/PostPageHeader.jsx
@@ -1,7 +1,7 @@
 import PaperInfo from '../PaperInfo';
-import * as S from './PaperHeader.style';
+import * as S from './PostPageHeader.style';
 
-export default function PaperHeader({
+export default function PostPageHeader({
   recentMessages,
   name,
   messageCount,
